% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_to_extent.R
\name{bbox_to_extent}
\alias{bbox_to_extent}
\title{bbox_to_extent}
\usage{
bbox_to_extent(sf_bbox)
}
\arguments{
\item{sf_bbox}{an object of class bbox created by sf::st_bbox}
}
\value{
a raster extent representing the bbox
}
\description{
Simple Features Bounding Box to Raster Extent
}
\details{
Converts a simple features bounding box to a raster extent.
}
