% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_to_trimesh.R
\name{bbox_to_trimesh}
\alias{bbox_to_trimesh}
\title{bbox_to_trimesh}
\usage{
bbox_to_trimesh(bbox, n_tris)
}
\arguments{
\item{bbox}{an sf bbox object}

\item{n_tris}{a number of triangles to use in the mesh (approximate)}
}
\value{
an RTriangle triangulation object
}
\description{
Triangulate a bounding box into a mesh
}
\details{
Triangulate an sf bounding box into a mesh using RTriangle

A wrapper for sf_to_trimesh()
}
