% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_transform.R
\name{bbox_transform}
\alias{bbox_transform}
\title{bbox_transform}
\usage{
bbox_transform(bbox, crs)
}
\arguments{
\item{bbox}{an sf bbox object.}

\item{crs}{an sf crs expression. A string, an EPSG code, an object created
with sf::st_crs() @return a bounding box transformed to the coordinate
reference system defined by `crs`}
}
\description{
Transform the coordinate reference system of a bounding box
}
\details{
sf does not let you transform a bounding box without jumping through some hoops. this function wraps those up.
}
