% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_scale.R
\name{range_scale}
\alias{range_scale}
\title{range_scale}
\usage{
range_scale(a)
}
\arguments{
\item{a}{vector to be scaled}
}
\value{
a scaled on 0 - 1
}
\description{
Scale a column on 0 - 1
}
\details{
A simple scaling operation that is useful for moving coordinates from images space to texture coordinate space.
}
