% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_crop_bbox.R
\name{raster_crop_bbox}
\alias{raster_crop_bbox}
\title{raster_crop_bbox}
\usage{
raster_crop_bbox(a_raster, sf_bbox)
}
\arguments{
\item{a_raster}{a Raster from the 'raster' package}

\item{sf_bbox}{a object of class 'bbox' created by sf::st_bbox()}
}
\value{
a_raster cropped to the extent of sf_bbox
}
\description{
Crop Raster to Simple Features Bounding Box
}
\details{
The bounding box is converted to a raster::extent which is used to crop the
raster. The bounding box need not have the same CRS as the raster, it is
transformed to the raster's CRS automatically.
}
