% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_pal_colours.R
\name{vec_pal_colours}
\alias{vec_pal_colours}
\title{vec_pal_colours}
\usage{
vec_pal_colours(vec, palette_fn, n_cols, zero_index = FALSE)
}
\arguments{
\item{vec}{a numeric of vector of values to generate palette indexes for}

\item{palette_fn}{a function that will produce a vector of integer colour code
strings, when called: `palette_fn(n_cols)`}

\item{n_cols}{the number of colours to generate using the `palette_fn`.}

\item{zero_index}{should the colour indexes begin at 0, for use with 0 indexed languages like Javascript?}
}
\value{
A list contain two named vectors: 'colours' and 'indexes'.
}
\description{
Create colour scale and indexes for vector
}
\details{
Given a vector of numeric values, return a palette of size `n_cols`` and a
set of indexes into the palette of size equal to the numeric vector.
}
